pyproject.toml
[tool.poetry]
name = "class-based-views"
version = "0.1.0"
description = ""
authors = ["Peter Marsh <p.marsh@mac.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
Pillow = "^8.3.2"
django-ckeditor = "^6.1.0"
django-crispy-forms = "^1.13.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
